requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
sentence-transformers>=2.7.0
faiss-cpu>=1.9.0
groq>=0.9.0
streamlit>=1.35.0
numpy>=1.26.0